Deploy test (#90)
* Update README.md to enhance clarity and organization

- Added spacing for improved readability in the Key Features and Available Tools sections.
- Refined installation instructions with clearer formatting.
- Improved example code formatting for better presentation.
- Enhanced the contribution guidelines for better clarity.

* Enhance project structure and update dependencies

- Updated TypeScript configuration to include new paths and target settings.
- Refactored layout components to improve structure and maintainability.
- Introduced ClientLayout component for better organization of layout logic.
- Updated package dependencies, including major upgrades for React, Next.js, and Ant Design.
- Added new deployment scripts for easier integration with Wrangler.
- Improved styles and responsiveness across various components.
- Enhanced the FAQ collapse component to utilize Ant Design's new item structure.
- Added new middleware and configurations for better handling of requests in the application.

* Add .npmrc for legacy peer dependencies and update package dependencies

- Created a new .npmrc file to set legacy-peer-deps to true.
- Updated @monaco-editor/react to version 4.7.0.
- Upgraded eslint to version 9 and adjusted related dependencies in package-lock.json.
- Enhanced compatibility with newer Node.js versions in various dependencies.

* Update README.md for improved punctuation and clarity

* Fix wrangler deploy by adding [assets] directory config


---------
