+++ .gitignore
@@ -29,6 +29,9 @@ pnpm-lock.yaml
 .next/
 out/
 build
+
+# Wrangler local dev state (auto-generated by wrangler pages dev)
+.wrangler/
 dist
 dist-ssr
