+++ app/components/comparisonsComponent/comparisonsStyles.module.scss
@@ -9,16 +9,6 @@
   height: 100%;
 }
 
-// .testimonialCard {
-//   margin-top: 40px;
-//   background: linear-gradient(180deg, #000000 0%, rgba(102, 102, 102, 0) 100%);
-//   border-top: 1px solid #00da92;
-//   border-right: 1px solid #00da92;
-//   border-left: 1px solid #00da92;
-//   border-bottom: 0;
-//   border-radius: 20px;
-// }
-
 .ctaCardGridBg {
   background: linear-gradient(
     180deg,
@@ -34,32 +24,30 @@
     inset 0 1px 0 rgba(255, 255, 255, 0.9);
 }
 
-.collapse {
-  :global(.ant-collapse) {
-    background-color: transparent !important;
+.collapse :global(.ant-collapse) {
+  background-color: transparent !important;
+}
 
-    :global(.ant-collapse-expand-icon) {
-      color: white !important;
-      order: 2;
-    }
+.collapse :global(.ant-collapse .ant-collapse-expand-icon) {
+  color: white !important;
+  order: 2;
+}
 
-    @media (min-width: 1024px) {
-      :global(.ant-collapse-header) {
-        padding: 0px 0px 16px 0px !important;
-      }
-    }
-
-    @media (max-width: 564px) {
-      :global(.ant-collapse-header) {
-        padding: 16px 0px 16px 0px !important;
-        align-items: center !important;
-      }
-    }
-
-    :global(.ant-collapse-content-box) {
-      padding: 0px 0px 20px 0px !important;
-    }
+@media (min-width: 1024px) {
+  .collapse :global(.ant-collapse .ant-collapse-header) {
+    padding: 0px 0px 16px 0px !important;
   }
+}
+
+@media (max-width: 564px) {
+  .collapse :global(.ant-collapse .ant-collapse-header) {
+    padding: 16px 0px 16px 0px !important;
+    align-items: center !important;
+  }
+}
+
+.collapse :global(.ant-collapse .ant-collapse-content-box) {
+  padding: 0px 0px 20px 0px !important;
 }
 
 .borderBottom {
