+++ app/components/headerMenuDropdownComponents/productMenuComponent.tsx
@@ -1,3 +1,4 @@
+'use client';
 import Image from 'next/image';
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
