+++ app/components/headerMenuDropdownComponents/resourceMenuComponent.tsx
@@ -1,3 +1,4 @@
+'use client';
 import React from 'react';
 import { usePathname } from 'next/navigation';
 import Link from 'next/link';
