+++ app/components/layout/Header/headerStyles.module.scss
@@ -1,43 +1,33 @@
 @import "../../../styles/variables";
 
-.accordion {
-  :global {
-    .ant-collapse {
-      .ant-collapse-item {
-        .ant-collapse-header {
-          padding: 0 !important;
-        }
-      }
-    }
+.accordion :global(.ant-collapse .ant-collapse-item .ant-collapse-header) {
+  padding: 0 !important;
+}
+
+.collapse :global(.ant-collapse) {
+  background-color: transparent !important;
+}
+
+.collapse :global(.ant-collapse .ant-collapse-expand-icon) {
+  color: white !important;
+  order: 2;
+}
+
+@media (min-width: 1024px) {
+  .collapse :global(.ant-collapse .ant-collapse-header) {
+    padding: 16px 0px 16px 0px !important;
   }
 }
 
-.collapse {
-  :global(.ant-collapse) {
-    background-color: transparent !important;
-
-    :global(.ant-collapse-expand-icon) {
-      color: white !important;
-      order: 2;
-    }
-
-    @media (min-width: 1024px) {
-      :global(.ant-collapse-header) {
-        padding: 16px 0px 16px 0px !important;
-      }
-    }
-
-    @media (max-width: 564px) {
-      :global(.ant-collapse-header) {
-        padding: 0px !important;
-        align-items: center !important;
-      }
-    }
-
-    :global(.ant-collapse-content-box) {
-      padding: 0px 0px 20px 0px !important;
-    }
+@media (max-width: 564px) {
+  .collapse :global(.ant-collapse .ant-collapse-header) {
+    padding: 0px !important;
+    align-items: center !important;
   }
+}
+
+.collapse :global(.ant-collapse .ant-collapse-content-box) {
+  padding: 0px 0px 20px 0px !important;
 }
 
 @keyframes blink {
@@ -71,12 +61,10 @@
     width: 4px !important;
   }
 
-  /* Track */
   &::-webkit-scrollbar-track {
     background: transparent;
   }
 
-  /* Handle */
   &::-webkit-scrollbar-thumb {
     background: $secondary;
     border-radius: 15px;
@@ -99,15 +87,15 @@
   &:hover {
     background: rgba(255, 255, 255, 0.16);
   }
+}
 
-  :global([data-theme="light"]) & {
-    background: rgba(17, 24, 39, 0.06);
-    border-color: rgba(17, 24, 39, 0.2);
-    color: #111827;
+:global([data-theme="light"]) .themeToggle {
+  background: rgba(17, 24, 39, 0.06);
+  border-color: rgba(17, 24, 39, 0.2);
+  color: #111827;
 
-    &:hover {
-      background: rgba(17, 24, 39, 0.12);
-    }
+  &:hover {
+    background: rgba(17, 24, 39, 0.12);
   }
 }
