+++ package.json
@@ -32,37 +32,41 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "deploy": "wrangler pages deploy out",
+    "build:deploy": "next build && wrangler pages deploy out",
+    "preview": "next build && wrangler pages dev out"
   },
   "dependencies": {
+    "@ant-design/cssinjs": "^2.1.2",
+    "@ant-design/nextjs-registry": "^1.3.0",
     "@betterbugs/web-sdk": "0.0.29",
     "@faker-js/faker": "9.4.0",
-    "@monaco-editor/react": "4.6.0",
+    "@monaco-editor/react": "4.7.0",
     "@paddle/paddle-js": "^1.6.2",
     "@phosphor-icons/react": "^2.1.10",
     "@types/turndown": "5.0.5",
     "animejs": "3.2.2",
-    "antd": "5.16.2",
+    "antd": "6.4.3",
     "bcryptjs": "^2.4.3",
     "class-validator": "^0.15.1",
     "curlconverter": "^4.12.0",
     "dayjs": "^1.11.19",
-    "framer-motion": "11.2.6",
+    "framer-motion": "12.40.0",
     "gleap": "^15.1.8",
     "javascript-obfuscator": "4.1.1",
     "js-user-country": "1.0.1",
     "lodash.debounce": "4.0.8",
     "marked": "15.0.6",
-    "next": "14.2.1",
-    "next-pwa": "5.6.0",
+    "next": "16.2.6",
     "qrcode.react": "^4.2.0",
-    "react": "18.2.0",
+    "react": "19.2.6",
     "react-animated-cursor": "2.11.2",
     "react-content-loader": "7.0.0",
-    "react-cookie": "^8.0.1",
-    "react-dom": "18.2.0",
+    "react-cookie": "^8.1.2",
+    "react-dom": "19.2.6",
     "react-hook-form": "7.51.5",
-    "react-responsive": "10.0.0",
+    "react-responsive": "10.0.1",
     "react-toastify": "11.0.3",
     "sass": "1.75.0",
     "turndown": "7.2.0"
@@ -77,13 +81,14 @@
     "@types/bcryptjs": "^2.4.6",
     "@types/lodash.debounce": "^4.0.9",
     "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "eslint": "^8",
-    "eslint-config-next": "14.2.1",
+    "@types/react": "^19",
+    "@types/react-dom": "^19",
+    "eslint": "^9",
+    "eslint-config-next": "16.2.6",
     "postcss": "^8",
     "semantic-release": "^23.0.8",
     "tailwindcss": "^3.4.1",
-    "typescript": "^5"
+    "typescript": "^5",
+    "wrangler": "^4.0.0"
   }
 }
