+++ wrangler.toml
@@ -1,0 +1,6 @@
+name = "dev-tools"
+pages_build_output_dir = "out"
+compatibility_date = "2025-01-01"
+
+[assets]
+directory = "./out"
